requirements.txt
commonmark==0.9.1
psycopg2-binary==2.9.3
Pygments==2.12.0
rich==12.3.0
